fixed message length

+++ CommunicationLibrary/CommunicationLibrary/RawMessageProcessing/RawMessageSender.cs
@@ -15,7 +15,7 @@ namespace CommunicationLibrary.RawMessageProcessing
         public void SendMessage(string messageString)
         {
             byte[] messageBytes = Encoding.UTF8.GetBytes(messageString);
-            byte[] messageLengthBytes = BitConverter.GetBytes((ushort)messageBytes.Length);
+            byte[] messageLengthBytes = BitConverter.GetBytes(messageBytes.Length);
             if (!BitConverter.IsLittleEndian) Array.Reverse(messageLengthBytes);
             byte[] allBytes = new byte[messageBytes.Length + messageLengthBytes.Length];
             for (int i = 0; i < messageLengthBytes.Length; i++)
